requirements.txt
python-telegram-bot==13.7
openai==0.28.0
python-dotenv==0.19.2
logging
